Clean code test_network_ops2

+++ test/test_network_ops/test_stack.py
@@ -19,6 +19,7 @@ import torch_npu
 from torch_npu.testing.testcase import TestCase, run_tests
 from torch_npu.testing.common_utils import create_common_tensor
 
+
 class TestStack(TestCase):
     def cpu_op_exec(self, input1, input2, dim):
         cpu_output = torch.stack((input1, input2), dim)
@@ -31,18 +32,18 @@ class TestStack(TestCase):
         output = output.numpy()
         return output
 
-    def cpu_op_exec_out(self, input1, input2, dim, input3):    
+    def cpu_op_exec_out(self, input1, input2, dim, input3):
         torch.stack((input1, input2), dim, out=input3)
         output = input3.numpy()
         return output
 
-    def npu_op_exec_out(self, input1, input2, dim, input3):    
+    def npu_op_exec_out(self, input1, input2, dim, input3):
         torch.stack((input1, input2), dim, out=input3)
         output = input3.to("cpu")
         output = output.numpy()
         return output
-    
-    def npu_output_size(self, inputs, dim = 0):
+
+    def npu_output_size(self, inputs, dim=0):
         shape = []
         for i in range(dim):
             shape.append(inputs[0].size(i))
@@ -56,9 +57,9 @@ class TestStack(TestCase):
         for item in shape_format:
             cpu_input1, npu_input1 = create_common_tensor(item[0], 0, 100)
             cpu_input2, npu_input2 = create_common_tensor(item[0], 0, 100)
-            shape = self.npu_output_size([npu_input1,npu_input2], item[1])
-            npu_input3 = torch.ones(shape, dtype = cpu_input1.dtype).npu()
-            cpu_input3 = torch.ones(shape, dtype = cpu_input1.dtype)
+            shape = self.npu_output_size([npu_input1, npu_input2], item[1])
+            npu_input3 = torch.ones(shape, dtype=cpu_input1.dtype).npu()
+            cpu_input3 = torch.ones(shape, dtype=cpu_input1.dtype)
             if cpu_input1.dtype == torch.float16:
                 cpu_input1 = cpu_input1.to(torch.float32)
                 cpu_input2 = cpu_input2.to(torch.float32)
@@ -138,20 +139,21 @@ class TestStack(TestCase):
             output = torch.stack((input1, input1, input1, input1, input1, input1, input1, input1, input1))
             return output.numpy()
 
-        def npu_op_exec(input1):        
+        def npu_op_exec(input1):
             output = torch.stack((input1, input1, input1, input1, input1, input1, input1, input1, input1))
             output = output.to("cpu")
             return output.numpy()
         shape_format = [
-                [[np.int32, 0, ()]],
-                [[np.float32, 0, ()]],
-                [[np.float16, 0, ()]],
+            [[np.int32, 0, ()]],
+            [[np.float32, 0, ()]],
+            [[np.float16, 0, ()]],
         ]
         for item in shape_format:
             cpu_input1, npu_input1 = create_common_tensor(item[0], 1, 100)
             cpu_output = cpu_op_exec(cpu_input1)
             npu_output = npu_op_exec(npu_input1)
             self.assertRtolEqual(cpu_output, npu_output)
+
 
 if __name__ == "__main__":
     run_tests()
